requirements.txt
# Django Core
Django>=5.2
django-allauth>=0.61.0
python-decouple>=3.8
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
whitenoise>=6.6.0

# OpenAI
openai>=1.0.0
python-dotenv>=1.0.0

# Image processing
Pillow>=10.0.0

# Database
dj-database-url>=2.1.0

# Embeddings
sentence-transformers>=3.0.0
torch>=2.0.0
